Add some modifications to files

+++ ProjectName.Application/DTO/AttendanceDTO.cs
@@ -8,7 +8,7 @@ namespace UMS.Application.DTO
 {
     public class AttendanceDTO
     {
-        public int StudentCourseId { get; set; }
+        public long StudentCourseId { get; set; }
         public bool AttendanceStatus { get; set; }
         public DateTime Date { get; set; }
     }
